+++ tests/PSDocs.Tests/PSDocs.BlockQuote.Tests.ps1
@@ -3,9 +3,7 @@
 #
 
 [CmdletBinding()]
-param (
-
-)
+param ()
 
 # Setup error handling
 $ErrorActionPreference = 'Stop';
@@ -15,6 +13,7 @@ Set-StrictMode -Version latest;
 $rootPath = $PWD;
 Import-Module (Join-Path -Path $rootPath -ChildPath out/modules/PSDocs) -Force;
 $here = (Resolve-Path $PSScriptRoot).Path;
+$nl = [System.Environment]::NewLine;
 
 Describe 'PSDocs -- BlockQuote keyword' -Tag BlockQuote {
     $docFilePath = Join-Path -Path $here -ChildPath 'FromFile.Keyword.Doc.ps1';
@@ -30,19 +29,19 @@ Describe 'PSDocs -- BlockQuote keyword' -Tag BlockQuote {
         }
         It 'Should handle single line input' {
             $result = Invoke-PSDocument @invokeParams -Name 'BlockQuoteSingleMarkdown';
-            $result | Should -Match '\> This is a single line';
+            $result | Should -Match "Begin$($nl)$($nl)\> This is a single line$($nl)$($nl)End";
         }
         It 'Should handle multiline input' {
             $result = Invoke-PSDocument @invokeParams -Name 'BlockQuoteMultiMarkdown';
-            $result | Should -Match '\> This is the first line.(\r|\n|\r\n)\> This is the second line.';
+            $result | Should -Match "Begin$($nl)$($nl)\> This is the first line.$($nl)\> This is the second line.$($nl)$($nl)End";
         }
         It 'Should add title' {
             $result = Invoke-PSDocument @invokeParams -Name 'BlockQuoteTitleMarkdown';
-            $result | Should -Match '\> Test(\r|\n|\r\n)\> This is a single block quote';
+            $result | Should -Match "Begin$($nl)$($nl)\> Test$($nl)\> This is a single block quote$($nl)$($nl)End";
         }
         It 'Should add info' {
             $result = Invoke-PSDocument @invokeParams -Name 'BlockQuoteInfoMarkdown';
-            $result | Should -Match '\> \[!TIP\](\r|\n|\r\n)> This is a single block quote';
+            $result | Should -Match "Begin$($nl)$($nl)\> \[!TIP\]$($nl)\> This is a single block quote$($nl)$($nl)End";
         }
     }
 }